move the test_capitals sanity test to the test_grpo_trainer.py
